requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
importlib-metadata==3.10.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
python-dotenv==0.17.0
SQLAlchemy==1.4.7
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.1
